pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "videolizer"
version = "0.1.0"
description = "Media Remix Engine for MofaClaw - renders premium video from VideoPlan"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
full = [
    "google-genai>=1.0.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=4.0.0",
    "torch>=2.5.0",
    "torchaudio>=2.5.0",
    "chatterbox-tts",
    "resemble-perth",
    "pydub>=0.25.0",
    "moviepy>=1.0.3",
    "Pillow>=11.0.0",
    "opencv-python-headless>=4.10.0",
    "requests>=2.32.0",
    "openai-whisper",
    "yt-dlp>=2025.1.0",
]

[project.scripts]
videolizer = "videolizer.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["videolizer"]
